Finish weapon refactoring to implement on hit effects
