+++ Assets/Scripts/Combat/WeaponConfig.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4c7e6afd8ebdd9d779b2a9b267a066e0
+guid: f69f5e2012e1529bca0f782252bbb47d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
